lambda/sync-redis/requirements.txt
asyncio==4.0.0
asyncpg==0.30.0
boto3==1.40.76
botocore==1.40.76
jmespath==1.0.1
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.0.1
s3transfer==0.14.0
six==1.17.0
urllib3==2.5.0
